Bug 1222640 - Fix truncated Simulator return values. r=jandem
UltraBlame original commit: 5d7f3436f9eceb1eb3b44c113ce31fa43159acc5

+++ js/src/jit/arm/Simulator-arm.h
@@ -181,7 +181,7 @@ class Simulator
     void execute();
 
     
-    int64_t call(uint8_t* entry, int argument_count, ...);
+    int32_t call(uint8_t* entry, int argument_count, ...);
 
     
     void setLastDebuggerInput(char* input);
